Create first task.sh
Write a bash script that searches for patterns in files located in a folder. The script should be 
called with two input parameter arguments. Make sure that the arguments are provided before 
proceeding. The first parameter should be a path to a folder and the second parameter should 
be a string pattern.